fix: recursively unpack nested archives
