+++ rocks-lib/src/operations/fetch.rs
@@ -1,17 +1,12 @@
 use bon::Builder;
-use flate2::read::GzDecoder;
 use git2::build::RepoBuilder;
 use git2::FetchOptions;
-use itertools::Itertools;
 use ssri::Integrity;
 use std::fs::File;
 use std::io;
-use std::io::BufReader;
 use std::io::Cursor;
 use std::io::Read;
-use std::io::Seek;
 use std::path::Path;
-use std::path::PathBuf;
 use thiserror::Error;
 
 use crate::config::Config;
@@ -24,6 +19,7 @@ use crate::progress::ProgressBar;
 use crate::rockspec::Rockspec;
 
 use super::DownloadSrcRockError;
+use super::UnpackError;
 
 /// A rocks package source fetcher, providing fine-grained control
 /// over how a package should be fetched.
@@ -162,7 +158,7 @@ async fn do_fetch_src<R: Rockspec>(fetch: &FetchSrc<'_, R>) -> Result<Integrity,
                 .unwrap_or(url.to_string());
             let cursor = Cursor::new(response);
             let mime_type = infer::get(cursor.get_ref()).map(|file_type| file_type.mime_type());
-            unpack(
+            operations::unpack::unpack(
                 mime_type,
                 cursor,
                 rock_source.unpack_dir.is_none(),
@@ -197,7 +193,7 @@ async fn do_fetch_src<R: Rockspec>(fetch: &FetchSrc<'_, R>) -> Result<Integrity,
                     .map(|os_str| os_str.to_string_lossy())
                     .unwrap_or(path.to_string_lossy())
                     .to_string();
-                unpack(
+                operations::unpack::unpack(
                     mime_type,
                     file,
                     rock_source.unpack_dir.is_none(),
@@ -226,124 +222,14 @@ async fn do_fetch_src_rock(fetch: FetchSrcRock<'_>) -> Result<Integrity, FetchSr
     let integrity = src_rock.bytes.hash()?;
     let cursor = Cursor::new(src_rock.bytes);
     let mime_type = infer::get(cursor.get_ref()).map(|file_type| file_type.mime_type());
-    unpack(
+    operations::unpack::unpack(
         mime_type,
         cursor,
-        false,
+        true,
         src_rock.file_name,
         dest_dir,
         progress,
     )
     .await?;
     Ok(integrity)
-}
-
-fn is_single_directory<R: Read + Seek + Send>(reader: R) -> io::Result<bool> {
-    let tar = GzDecoder::new(reader);
-    let mut archive = tar::Archive::new(tar);
-
-    let entries: Vec<_> = archive
-        .entries()?
-        .filter_map(|entry| {
-            if entry.as_ref().ok()?.path().ok()?.file_name()? != "pax_global_header" {
-                Some(entry)
-            } else {
-                None
-            }
-        })
-        .try_collect()?;
-
-    let directory: PathBuf = entries
-        .first()
-        .unwrap()
-        .path()?
-        .components()
-        .take(1)
-        .collect();
-
-    Ok(entries.into_iter().all(|entry| {
-        entry
-            .path()
-            .unwrap()
-            .to_str()
-            .unwrap()
-            .starts_with(directory.to_str().unwrap())
-    }))
-}
-
-#[derive(Error, Debug)]
-pub enum UnpackError {
-    #[error(transparent)]
-    Zip(#[from] zip::result::ZipError),
-    #[error(transparent)]
-    Io(#[from] io::Error),
-    #[error("source returned HTML - it may have been moved or deleted")]
-    SourceMovedOrDeleted,
-    #[error("rockspec source has unsupported file type {0}")]
-    UnsupportedFileType(String),
-    #[error("could not determine mimetype of rockspec source")]
-    UnknownMimeType,
-}
-
-async fn unpack<R: Read + Seek + Send>(
-    mime_type: Option<&str>,
-    reader: R,
-    auto_find_lua_sources: bool,
-    file_name: String,
-    dest_dir: &Path,
-    progress: &Progress<ProgressBar>,
-) -> Result<(), UnpackError> {
-    progress.map(|p| p.set_message(format!("📦 Unpacking {}", file_name)));
-
-    match mime_type {
-        Some("application/zip") => {
-            let mut archive = zip::ZipArchive::new(reader)?;
-            archive.extract(dest_dir)?;
-        }
-        Some("application/x-tar") => {
-            let mut archive = tar::Archive::new(reader);
-            archive.unpack(dest_dir)?;
-        }
-        Some("application/gzip") => {
-            let mut bufreader = BufReader::new(reader);
-
-            let extract_subdirectory =
-                auto_find_lua_sources && is_single_directory(&mut bufreader)?;
-
-            bufreader.rewind()?;
-            let tar = GzDecoder::new(bufreader);
-            let mut archive = tar::Archive::new(tar);
-
-            if extract_subdirectory {
-                archive.entries()?.try_for_each(|entry| {
-                    let mut entry = entry?;
-
-                    let path: PathBuf = entry.path()?.components().skip(1).collect();
-                    if path.components().count() > 0 {
-                        let dest = dest_dir.join(path);
-                        std::fs::create_dir_all(dest.parent().unwrap())?;
-                        entry.unpack(dest)?;
-                    }
-
-                    Ok::<_, io::Error>(())
-                })?;
-            } else {
-                archive.entries()?.try_for_each(|entry| {
-                    entry?.unpack_in(dest_dir)?;
-                    Ok::<_, io::Error>(())
-                })?;
-            }
-        }
-        Some("text/html") => {
-            return Err(UnpackError::SourceMovedOrDeleted);
-        }
-        Some(other) => {
-            return Err(UnpackError::UnsupportedFileType(other.to_string()));
-        }
-        None => {
-            return Err(UnpackError::UnknownMimeType);
-        }
-    }
-
-    Ok(())
 }
